backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
